laboratories/laboratory_4/1019142922/services/api-gateway/pyproject.toml
[project]
name = "api-gateway"
version = "0.1.0"
description = "A simple API Gateway"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.11.4",
    "pyjwt>=2.10.1",
]
